V1/requirements.txt
azure-core==1.30.1
azure-identity==1.15.0
azure-storage-blob==12.19.1
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
dnspython==2.6.1
Flask==3.0.2
fonttools==4.50.0
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
msal==1.28.0
msal-extensions==1.1.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.2.0
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
pymongo==4.6.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
soupsieve==2.5
threadpoolctl==3.3.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
